Issue#226: Fixed after merge conflict.
Returned description cooperation_id in openapi.
PollsApiImpl moved to PollApiImpl.
